backend/requirements.txt
asgiref==3.7.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
cryptography==41.0.7
daphne==4.0.0
defusedxml==0.8.0rc2
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
filetype==1.2.0
idna==3.6
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.8.6
python-dotenv
pytz==2023.3.post1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0